refactor: flatten overengineered factories and remove dead indirection
- Flatten the two-layer Supabase auth factory: inline
  createSupabaseAuthCoordinator and createSupabaseSessionStorage (each
  called exactly once) into createHostAuthCoordinator, and delete the
  re-export-only src/auth/createHostAuthCoordinator.ts
- Merge the follow-up, bypass, and approval sub-factories (each called
  once in production) into createProgressViewCommandHandlers; fold their
  tests into the existing ProgressViewCommandHandlers suite
- Delete dead barrels with zero importers: src/eventBus/index.ts,
  src/hosts/index.ts, src/agent/remote/index.ts
- Delete the src/latex/index.ts facade (85 lines re-exporting 30+
  symbols with only 4 real consumers); point those imports at the
  source modules
- Trim src/logger/index.ts to its one consumed export
- Drop the crypto.randomUUID fallback ceremony in walkthroughDialog;
  a module counter already guarantees unique aria ids
- Derive the proposal FileFields type via schema composition instead of
  a manual intersection
