+++ src/agent/implementations/flows/reflection/ReflectionServices.ts
@@ -3,13 +3,13 @@ import type { RoundFinalizedCallback } from '@agent/core/flows/BaseFlowServices'
 import type { OutputState } from '@agent/output/outputState';
 import type { LatexDiffManager } from '@agent/output/LatexDiffManager';
 import type { XmlOutputManager } from '@agent/output/XmlOutputManager';
+import type { LatexMediaManager } from '@latex/LatexMediaManager';
+import type { PromptBuilder } from '@utils/prompt';
 import type {
   BaseFlowContextInit,
   FlowParams,
 } from '@agent/core/flows/BaseFlowServices';
-import type { LatexMediaManager } from '@latex';
 import type { AgentFileLocation, WorkspaceFileLocation } from '@shared/schemas';
-import type { PromptBuilder } from '@utils/prompt';
 import type { TaskRunFileService } from '@utils/files';
 
 export interface ReflectionServices<
