--- src/auth/createHostAuthCoordinator.ts
@@ -1,4 +1,0 @@
-export {
-  createHostAuthCoordinator,
-  type HostAuthCoordinatorInit,
-} from './SupabaseAuthCoordinator';
