--- src/latex/index.ts
@@ -1,11 +1,0 @@
-// -----------------------------------------------------------------------------
-// Media/Image Handling
-// -----------------------------------------------------------------------------
-
-export { LatexMediaManager } from './LatexMediaManager';
-
-// -----------------------------------------------------------------------------
-// Core Utilities
-// -----------------------------------------------------------------------------
-
-export { getTeXCountStats } from './texcount';
